+++ CHANGELOG.md
@@ -16,6 +16,14 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Breaking
 
+## 2025-12-22
+
+- richat-shared-v7.2.0
+
+### Features
+
+- shared: add config loading ([#169](https://github.com/lamports-dev/richat/pull/169))
+
 ## 2025-12-08
 
 - richat-shared-v7.1.0